react_refresh: Rewrite main loop
This turned into an all-in-one commit:

 * Move the logic to the `visit_mut_*` functions from VisitMut.
 * Make atom key names more flexible to support e.g. #26

The largest change is the fact that we perform most of the work directly
in the visitor functions.  For applicable calls to `atom` (and friends)
we replace the `CallExpr` right then and there.  This makes it so that
we don't need to deal with `export` handling, or accidentally replace
entire expressions with erroneous cached atoms (See e.g. the failing
test in #36).  This also solves other open issues, like #26, #34, and #35.

Fixes #26
Fixes #34
Fixes #35